# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateStock:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
      store_id:
        type: integer
    type: object
...
  models.StockPrimaryKey:
    properties:
      store_id:
        type: integer
    type: object
...
  models.UpdateStock:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
      store_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /stock:
    get:
      consumes:
      - application/json
      description: Get List Stock
      operationId: get_list_stock
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Stock
      tags:
      - Stock
    post:
      consumes:
      - application/json
      description: Create Stock
      operationId: create_stock
      parameters:
      - description: CreateStockRequest
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/models.CreateStock'
      produces:
      - application/json
      responses:
        "201":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Stock
      tags:
      - Stock
  /stock/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Stock
      operationId: delete_stock
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: DeleteStockRequest
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/models.StockPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Stock
      tags:
      - Stock
    get:
      consumes:
      - application/json
      description: Get By ID Stock
      operationId: get_by_id_stock
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Stock
      tags:
      - Stock
    put:
      consumes:
      - application/json
      description: Update Stock
      operationId: update_stock
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateStockRequest
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStock'
      produces:
      - application/json
      responses:
        "202":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Stock
      tags:
      - Stock
...
swagger: "2.0"